move callable_store to Utils
